DS-913 Separate Non Common Functions (#731)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-913>**

## Description of Changes

This PR splits non common functions from the common folder info the
respective lambda functions. As well it removes a bit of remaining dead
code,
