+++ .github/workflows/code-quality.yml
@@ -36,7 +36,7 @@ jobs:
             -Dsonar.sources=application,deployment,infrastructure,scripts,test,build/docker
             -Dsonar.organization=nhsd-exeter
             -Dsonar.projectKey=uec-dos-int
-            -Dsonar.coverage.exclusions=tests/**,**/tests/**,deployment,infrastructure,application/dos_db_handler/**,test/**,scripts/**
+            -Dsonar.coverage.exclusions=tests/**,**/tests/**,deployment,infrastructure,application/dos_db_handler/**,test/**,scripts/**,application/conftest.py
             -Dsonar.python.coverage.reportPaths=coverage.xml
             -Dsonar.python.version=3.10
             -Dsonar.exclusions=application/**/tests/**
