+++ application/common/constants.py
@@ -1,22 +1,3 @@
-HIDDEN_OR_CLOSED_REPORT_ID = "HIDDEN_OR_CLOSED"
-UNMATCHED_PHARMACY_REPORT_ID = "UNMATCHED_PHARMACY"
-INVALID_POSTCODE_REPORT_ID = "INVALID_POSTCODE"
-INVALID_OPEN_TIMES_REPORT_ID = "INVALID_OPEN_TIMES"
-DLQ_HANDLER_REPORT_ID = "CR_DLQ_HANDLER_RECEIVED_EVENT"
-FIFO_DLQ_HANDLER_REPORT_ID = "FIFO_DLQ_HANDLER_RECEIVED_EVENT"
-UNMATCHED_SERVICE_TYPE_REPORT_ID = "UNMATCHED_SERVICE_TYPE"
-GENERIC_BANK_HOLIDAY_REPORT_ID = "GENERIC_BANK_HOLIDAY"
-GENERIC_CHANGE_EVENT_ERROR_REPORT_ID = "GENERIC_CHANGE_EVENT_ERROR"
-SERVICE_UPDATE_REPORT_ID = "SERVICE_UPDATE"
-BLANK_STANDARD_OPENINGS_REPORT_ID = "BLANK_STANDARD_OPENINGS"
-INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE_REPORT_ID = "INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE"
-UNEXPECTED_PHARMACY_PROFILING_REPORT_ID = "UNEXPECTED_PHARMACY_PROFILING"
-
-METRIC_REPORT_KEY_MAP = {
-    "InvalidPostcode": INVALID_POSTCODE_REPORT_ID,
-    "InvalidOpenTimes": INVALID_OPEN_TIMES_REPORT_ID,
-}
-
 PHARMACY_SERVICE_KEY = "PHARMACY"
 DENTIST_SERVICE_KEY = "DENTIST"
 PHARMACY_ORG_TYPE_ID = "PHA"
