+++ application/service_sync/tests/test_reporting.py
@@ -1,0 +1,206 @@
+from os import environ
+from unittest.mock import MagicMock, patch
+
+from aws_lambda_powertools.logging import Logger
+
+from application.conftest import dummy_dos_service
+from application.service_sync.reporting import (
+    BLANK_STANDARD_OPENINGS_REPORT_ID,
+    GENERIC_BANK_HOLIDAY_REPORT_ID,
+    GENERIC_CHANGE_EVENT_ERROR_REPORT_ID,
+    INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE_REPORT_ID,
+    INVALID_POSTCODE_REPORT_ID,
+    SERVICE_UPDATE_REPORT_ID,
+    log_blank_standard_opening_times,
+    log_incorrect_palliative_stockholder_type,
+    log_invalid_nhsuk_postcode,
+    log_service_updated,
+    log_service_with_generic_bank_holiday,
+    log_website_is_invalid,
+)
+from common.nhs import NHSEntity
+from common.opening_times import OpenPeriod
+
+
+@patch.object(Logger, "warning")
+def test_log_blank_standard_opening_times(mock_logger, change_event):
+    # Arrange
+    nhs_entity = NHSEntity(change_event)
+    dos_service = dummy_dos_service()
+    dos_service.region = "London"
+    # Act
+    log_blank_standard_opening_times(nhs_entity, dos_service)
+    # Assert
+    assert (
+        BLANK_STANDARD_OPENINGS_REPORT_ID == "BLANK_STANDARD_OPENINGS"
+    ), f"Report ID should be BLANK_STANDARD_OPENINGS but was {BLANK_STANDARD_OPENINGS_REPORT_ID}"
+    mock_logger.assert_called_with(
+        "NHS Service has matching DoS services but no given standard opening times.",
+        extra={
+            "report_key": BLANK_STANDARD_OPENINGS_REPORT_ID,
+            "nhsuk_odscode": nhs_entity.odscode,
+            "dos_service_id": dos_service.id,
+            "dos_service_uid": dos_service.uid,
+            "dos_service_name": dos_service.name,
+            "nhsuk_service_status": nhs_entity.org_status,
+            "nhsuk_service_type": nhs_entity.org_type,
+            "nhsuk_sector": nhs_entity.org_sub_type,
+            "dos_service_status": dos_service.statusid,
+            "dos_service_type": dos_service.typeid,
+            "dos_service_type_name": dos_service.servicename,
+            "dos_region": dos_service.region,
+        },
+    )
+
+
+@patch.object(Logger, "warning")
+def test_log_invalid_nhsuk_postcode(mock_logger):
+    # Arrange
+    environ["ENV"] = "dev"
+    county = "county"
+    city = "city"
+    nhs_entity = NHSEntity(
+        {"Address1": "address1", "Address2": "address2", "Address3": "address3", "City": city, "County": county},
+    )
+    nhs_entity.odscode = "SLC4X"
+    nhs_entity.org_name = "OrganisationName"
+    nhs_entity.org_type_id = "PHA"
+    nhs_entity.org_status = "OrganisationStatus"
+    nhs_entity.org_sub_type = "OrganisationSubType"
+    nhs_entity.postcode = "MK2 XXX"
+
+    dos_service = dummy_dos_service()
+    # Act
+    log_invalid_nhsuk_postcode(nhs_entity, dos_service)
+    # Assert
+    assert (
+        INVALID_POSTCODE_REPORT_ID == "INVALID_POSTCODE"
+    ), f"Log ID should be INVALID_POSTCODE but was {INVALID_POSTCODE_REPORT_ID}"
+    mock_logger.assert_called_with(
+        f"NHS entity '{nhs_entity.odscode}' postcode '{nhs_entity.postcode}' is not a valid DoS postcode!",
+        extra={
+            "report_key": INVALID_POSTCODE_REPORT_ID,
+            "nhsuk_odscode": nhs_entity.odscode,
+            "nhsuk_organisation_name": nhs_entity.org_name,
+            "nhsuk_organisation_type": nhs_entity.org_type,
+            "nhsuk_organisation_subtype": nhs_entity.org_sub_type,
+            "nhsuk_address1": nhs_entity.address_lines[0],
+            "nhsuk_address2": nhs_entity.address_lines[1],
+            "nhsuk_address3": nhs_entity.address_lines[2],
+            "nhsuk_city": city,
+            "nhsuk_postcode": nhs_entity.postcode,
+            "nhsuk_county": county,
+            "validation_error_reason": "Postcode not valid/found on DoS",
+            "dos_service": dos_service.uid,
+            "dos_service_type_name": dos_service.servicename,
+        },
+    )
+    # Clean up
+    del environ["ENV"]
+
+
+@patch.object(Logger, "warning")
+def test_log_service_with_generic_bank_holiday(mock_logger):
+    # Arrange
+    nhs_entity = NHSEntity({})
+    nhs_entity.odscode = "SLC4X"
+    nhs_entity.org_name = "OrganisationName"
+    dos_service = dummy_dos_service()
+    open_periods = [OpenPeriod.from_string("08:00-13:00"), OpenPeriod.from_string("04:00-18:00")]
+    dos_service.standard_opening_times.generic_bankholiday = open_periods
+
+    # Act
+    log_service_with_generic_bank_holiday(nhs_entity, dos_service)
+    # Assert
+    mock_logger.assert_called_with(
+        f"DoS Service uid={dos_service.uid} has a generic BankHoliday Standard opening time set in DoS",
+        extra={
+            "report_key": GENERIC_BANK_HOLIDAY_REPORT_ID,
+            "nhsuk_odscode": nhs_entity.odscode,
+            "nhsuk_organisation_name": nhs_entity.org_name,
+            "dos_service_uid": dos_service.uid,
+            "dos_service_name": dos_service.name,
+            "dos_service_type_id": dos_service.typeid,
+            "bank_holiday_opening_times": OpenPeriod.list_string(open_periods),
+            "nhsuk_parentorg": nhs_entity.parent_org_name,
+            "dos_service_type_name": dos_service.servicename,
+        },
+    )
+
+
+@patch.object(Logger, "warning")
+def test_log_website_is_invalid(mock_logger: MagicMock):
+    # Arrange
+    nhs_entity = NHSEntity({})
+    nhs_entity.website = nhs_website = "http://www.google.com"
+    # Act
+    log_website_is_invalid(nhs_entity, nhs_website)
+    # Assert
+    mock_logger.assert_called_with(
+        f"Website is not valid, {nhs_website=}",
+        extra={
+            "report_key": GENERIC_CHANGE_EVENT_ERROR_REPORT_ID,
+            "error_reason": "Website is not valid",
+            "error_info": f"NHSUK unedited website: '{nhs_entity.website}', NHSUK website='{nhs_website}'",
+            "nhs_unedited_website": nhs_entity.website,
+            "nhs_website": nhs_website,
+        },
+    )
+
+
+@patch.object(Logger, "warning")
+def test_log_service_updated(mock_logger: MagicMock):
+    # Arrange
+    action = "action"
+    data_field_modified = "data_field_modified"
+    new_value = "new_value"
+    previous_value = "previous_value"
+    service_name = "ServiceName"
+    service_uid = "1234567890"
+    type_id = "1"
+    # Act
+    log_service_updated(
+        action=action,
+        data_field_modified=data_field_modified,
+        new_value=new_value,
+        previous_value=previous_value,
+        service_name=service_name,
+        service_uid=service_uid,
+        type_id=type_id,
+    )
+    # Assert
+    assert SERVICE_UPDATE_REPORT_ID == "SERVICE_UPDATE"
+    mock_logger.assert_called_with(
+        "Service update complete",
+        extra={
+            "report_key": SERVICE_UPDATE_REPORT_ID,
+            "action": action,
+            "previous_value": previous_value,
+            "new_value": new_value,
+            "data_field_modified": data_field_modified,
+            "service_name": service_name,
+            "service_uid": service_uid,
+            "type_id": type_id,
+        },
+    )
+
+
+@patch.object(Logger, "warning")
+def test_log_incorrect_palliative_stockholder_type(mock_logger: MagicMock):
+    # Arrange
+    expected_dos_palliative_care = True
+    expected_nhsuk_palliative_care = False
+    dos_service = dummy_dos_service()
+    # Act
+    log_incorrect_palliative_stockholder_type(expected_nhsuk_palliative_care, expected_dos_palliative_care, dos_service)
+    # Assert
+    assert INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE_REPORT_ID == "INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE"
+    mock_logger.assert_called_with(
+        "Palliative care on wrong service type",
+        extra={
+            "report_key": INCORRECT_PALLIATIVE_STOCKHOLDER_TYPE_REPORT_ID,
+            "dos_palliative_care": expected_dos_palliative_care,
+            "nhsuk_palliative_care": expected_nhsuk_palliative_care,
+            "dos_service_type_name": dos_service.servicename,
+        },
+    )
