+++ application/service_sync/tests/test_s3.py
@@ -1,9 +1,9 @@
 from os import environ
 from unittest.mock import patch
 
-from application.common.s3 import put_content_to_s3
+from application.service_sync.s3 import put_content_to_s3
 
-FILE_PATH = "application.common.s3"
+FILE_PATH = "application.service_sync.s3"
 
 
 @patch(f"{FILE_PATH}.client")
@@ -17,7 +17,10 @@ def test_put_content_to_s3(mock_client):
     # Assert
     mock_client.assert_called_once_with("s3")
     mock_client.return_value.put_object.assert_called_once_with(
-        Body=content, Bucket=bucket_name, Key=s3_filename, ServerSideEncryption="AES256",
+        Body=content,
+        Bucket=bucket_name,
+        Key=s3_filename,
+        ServerSideEncryption="AES256",
     )
     # Cleanup
     del environ["SEND_EMAIL_BUCKET_NAME"]
