+++ application/test_resources/STANDARD_EVENT_WITH_STAFF.json
